+++ rdt/__init__.py
@@ -5,7 +5,7 @@
 
 __author__ = 'DataCebo, Inc.'
 __email__ = 'info@sdv.dev'
-__version__ = '1.6.1'
+__version__ = '1.7.0.dev3'
 
 
 import sys